+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial Python InputObject Database definition layer. See [wasppy/README.md](wasppy/README.md) for more information. This allows Python application to define, deserialize, and provide input diagnostics for application input.
 
 ### Changed
+- LSP IOStreamConnection to handle turning off and on standard output when needed
 - EDDI comment logic where subsequent comments committed staged components
 - TreeNodePool node_token_line now recursively finds first leaf if node is not already a leaf, instead of returning illegal line number (0)